refactor popup.js: DRY stat rendering, extract render-detail and sync modules
