+++ popup.js
@@ -17,6 +17,19 @@ import {
   setElementColor,
   clearElementColors
 } from "./dom.js"
+import {
+  OVER_MASTERY_NAMES, AETHERIAL_NAMES, PERPETUITY_NAMES,
+  formatModifier, formatPerpetuityBonus
+} from "./mastery.js"
+import { RARITY_LABELS, GAME_ELEMENT_NAMES } from "./game-data.js"
+import { handleDetailSync, hideSyncModal, confirmSync } from "./sync.js"
+import {
+  isCollectionType, isDatabaseDetailType, isWeaponOrSummonCollection,
+  toArray, extractItems, countItems,
+  getItemImageUrl, getArtifactLabels, getGridClass,
+  getCharacterModifiers, renderCharacterModifiers, getWeaponModifiers, renderWeaponModifiers,
+  renderPartyDetail, renderDatabaseDetail
+} from "./render-detail.js"
 
 // ==========================================
 // STATE
@@ -40,7 +53,6 @@ let brokenImageIndices = new Set() // Track items with broken images (persists a
 // Filter state
 let activeRarityFilters = new Set(['4']) // SSR by default
 let excludeLv1Items = true
-const RARITY_LABELS = { '4': 'SSR', '3': 'SR', '2': 'R' }
 
 // Age ticker state
 let ageTickerInterval = null
@@ -174,11 +186,11 @@ function initializeEventListeners() {
   document.getElementById('detailBack')?.addEventListener('click', hideDetailView)
   document.getElementById('detailCopy')?.addEventListener('click', handleDetailCopy)
   document.getElementById('detailImport')?.addEventListener('click', handleDetailImport)
-  document.getElementById('detailSync')?.addEventListener('click', handleDetailSync)
+  document.getElementById('detailSync')?.addEventListener('click', () => handleDetailSync(currentDetailDataType, showToast))
 
   // Sync modal buttons
   document.getElementById('cancelSync')?.addEventListener('click', hideSyncModal)
-  document.getElementById('confirmSync')?.addEventListener('click', confirmSync)
+  document.getElementById('confirmSync')?.addEventListener('click', () => confirmSync(currentDetailDataType, showToast))
   document.querySelector('.modal-backdrop')?.addEventListener('click', hideSyncModal)
 
   // Filter listeners
@@ -410,125 +422,8 @@ function hideDetailView() {
   }
 }
 
-/**
- * Count items in data
- */
-function countItems(dataType, data) {
-  const items = extractItems(dataType, data)
-  return items.length
-}
-
 // Checkmark SVG for checkboxes
 const CHECK_ICON = `<svg width="14" height="14" viewBox="0 0 14 14" fill="currentColor"><path fill-rule="evenodd" clip-rule="evenodd" d="M12.7139 4.04764C13.14 3.52854 13.0837 2.74594 12.5881 2.29964C12.0925 1.85335 11.3453 1.91237 10.9192 2.43147L5.28565 9.94404L3.02018 7.32366C2.55804 6.83959 1.80875 6.83959 1.34661 7.32366C0.884464 7.80772 0.884464 8.59255 1.34661 9.07662L4.50946 12.6369C4.9716 13.121 5.72089 13.121 6.18303 12.6369C6.2359 12.5816 6.28675 12.5271 6.33575 12.4674L12.7139 4.04764Z"/></svg>`
-
-
-/**
- * Get character modifiers (perpetuity ring status)
- */
-function getCharacterModifiers(item) {
-  const param = item.param || {}
-  return {
-    perpetuity: !!param.has_npcaugment_constant
-  }
-}
-
-/**
- * Render character modifier overlay (perpetuity ring)
- */
-function renderCharacterModifiers(item) {
-  const mods = getCharacterModifiers(item)
-
-  if (!mods.perpetuity) return ''
-
-  return `<div class="char-modifiers">
-    <img class="perpetuity-ring" src="icons/perpetuity/filled.svg" alt="Perpetuity Ring" title="Perpetuity Ring">
-  </div>`
-}
-
-// Map awakening form names to icon filenames
-const WEAPON_AWAKENING_ICONS = {
-  'Attack': 'weapon-atk',
-  'Defense': 'weapon-def',
-  'Multiattack': 'weapon-multi',
-  'Charge Attack': 'weapon-ca',
-  'Skill': 'weapon-skill',
-  'Healing': 'weapon-heal',
-  'Special': 'weapon-special'
-}
-
-/**
- * Get weapon modifiers (awakening, ax skill, befoulment)
- */
-function getWeaponModifiers(item) {
-  const param = item.param || {}
-  const odiant = param.odiant || {}
-  const isOdiant = odiant.is_odiant_weapon === true
-
-  return {
-    awakening: param.arousal?.is_arousal_weapon ? param.arousal : null,
-    axSkill: !isOdiant ? param.augment_skill_info?.[0] || null : null,
-    befoulment: isOdiant ? {
-      skill: param.augment_skill_info?.[0]?.[0] || null,
-      exorcismLevel: odiant.exorcision_level || 0,
-      maxExorcismLevel: odiant.max_exorcision_level || 5,
-      iconImage: param.augment_skill_icon_image?.[0] || null
-    } : null
-  }
-}
-
-/**
- * Render weapon modifier overlay (awakening, ax skill, befoulment icons)
- */
-function renderWeaponModifiers(item) {
-  const mods = getWeaponModifiers(item)
-  if (!mods.awakening && !mods.axSkill && !mods.befoulment) return ''
-
-  let html = '<div class="weapon-modifiers">'
-
-  if (mods.awakening) {
-    const iconName = WEAPON_AWAKENING_ICONS[mods.awakening.form_name] || 'weapon-atk'
-    html += `<img class="awakening-icon" src="${getImageUrl(`awakening/${iconName}.png`)}" alt="Awakening" title="${mods.awakening.form_name} Lv.${mods.awakening.level}">`
-  }
-
-  if (mods.axSkill) {
-    html += `<img class="ax-skill-icon" src="${getImageUrl('ax/atk.png')}" alt="AX Skill" title="AX Skill">`
-  }
-
-  if (mods.befoulment) {
-    const skill = mods.befoulment.skill
-    const exLevel = mods.befoulment.exorcismLevel
-    const maxLevel = mods.befoulment.maxExorcismLevel
-    const showValue = skill?.show_value || 'Befouled'
-    const iconImage = mods.befoulment.iconImage || 'ex_skill_def_down'
-    html += `<img class="befoulment-icon" src="${getImageUrl(`ax/${iconImage}.png`)}" alt="Befoulment" title="Befoulment: ${showValue} (Exorcism ${exLevel}/${maxLevel})">`
-  }
-
-  html += '</div>'
-  return html
-}
-
-/**
- * Check if a data type is a collection type (supports item selection)
- */
-function isCollectionType(dataType) {
-  return dataType.startsWith('collection_') || dataType.startsWith('list_') || dataType.startsWith('stash_') || dataType === 'character_stats'
-}
-
-/**
- * Check if a data type is a database detail type
- */
-function isDatabaseDetailType(dataType) {
-  return dataType.startsWith('detail_')
-}
-
-/**
- * Check if a data type is a weapon or summon collection (supports rarity filtering)
- */
-function isWeaponOrSummonCollection(dataType) {
-  return dataType === 'collection_weapon' || dataType === 'collection_summon' ||
-         dataType === 'list_weapon' || dataType === 'list_summon' ||
-         dataType.startsWith('stash_weapon') || dataType.startsWith('stash_summon')
-}
 
 /**
  * Initialize filter dropdown and checkbox listeners
@@ -823,204 +718,6 @@ function updateSelectionCount() {
 /**
  * Convert object or array to array (handles GBF's inconsistent data formats)
  */
-function toArray(data) {
-  if (!data) return []
-  if (Array.isArray(data)) return data
-  return Object.values(data)
-}
-
-/**
- * Render party detail with sections for job, characters, weapons, summons, accessories
- */
-function renderPartyDetail(container, data) {
-  // Characters are at deck.npc, weapons/summons are at deck.pc
-  const deck = data.deck || {}
-  const pc = deck.pc || {}
-  const job = pc.job
-  const characters = toArray(deck.npc).filter(Boolean)
-  const weapons = toArray(pc.weapons).filter(Boolean)
-  const summons = toArray(pc.summons).filter(Boolean)
-
-  // Accessories: familiar_id (manabelly/manatura) and shield_id
-  const accessoryIds = [pc.familiar_id, pc.shield_id].filter(Boolean)
-
-  let html = ''
-
-  // Job section
-  if (job?.master?.id) {
-    const jobId = job.master.id
-    const jobName = job.master.name || 'Job'
-    // Use Gran (a) by default, could detect from data if available
-    const jobImageUrl = getImageUrl(`job-wide/${jobId}_a.jpg`)
-    html += `
-      <div class="party-section">
-        <h3 class="party-section-title">Job</h3>
-        <div class="wide-item">
-          <img src="${jobImageUrl}" alt="${jobName}">
-        </div>
-      </div>
-    `
-  }
-
-  // Characters section
-  if (characters.length > 0) {
-    html += `
-      <div class="party-section">
-        <h3 class="party-section-title">Characters</h3>
-        <div class="item-grid characters">
-          ${characters.map(item => {
-            const id = item.master?.id || item.param?.id || item.id
-            const imageUrl = getImageUrl(`character-grid/${id}_01.jpg`)
-            return `
-              <div class="grid-item">
-                <img src="${imageUrl}" alt="">
-              </div>
-            `
-          }).join('')}
-        </div>
-      </div>
-    `
-  }
-
-  // Weapons section
-  if (weapons.length > 0) {
-    html += `
-      <div class="party-section">
-        <h3 class="party-section-title">Weapons</h3>
-        <div class="item-grid weapons">
-          ${weapons.map(item => {
-            const id = item.master?.id || item.param?.id || item.id
-            const imageUrl = getImageUrl(`weapon-grid/${id}.jpg`)
-            return `
-              <div class="grid-item">
-                <img src="${imageUrl}" alt="">
-              </div>
-            `
-          }).join('')}
-        </div>
-      </div>
-    `
-  }
-
-  // Summons section
-  if (summons.length > 0) {
-    html += `
-      <div class="party-section">
-        <h3 class="party-section-title">Summons</h3>
-        <div class="item-grid summons">
-          ${summons.map(item => {
-            const id = item.master?.id || item.param?.id || item.id
-            const imageUrl = getImageUrl(`summon-wide/${id}.jpg`)
-            return `
-              <div class="grid-item">
-                <img src="${imageUrl}" alt="">
-              </div>
-            `
-          }).join('')}
-        </div>
-      </div>
-    `
-  }
-
-  // Accessories section (manabelly, manatura, shields)
-  if (accessoryIds.length > 0) {
-    html += `
-      <div class="party-section">
-        <h3 class="party-section-title">Accessories</h3>
-        <div class="item-grid accessories">
-          ${accessoryIds.map(id => {
-            const imageUrl = getImageUrl(`accessory-square/${id}.jpg`)
-            return `
-              <div class="grid-item">
-                <img src="${imageUrl}" alt="">
-              </div>
-            `
-          }).join('')}
-        </div>
-      </div>
-    `
-  }
-
-  container.innerHTML = html || '<p class="cache-empty">No party data</p>'
-}
-
-// Over Mastery (ring) modifier ID to name mapping
-const OVER_MASTERY_NAMES = {
-  1: 'ATK', 2: 'HP', 3: 'Debuff Success', 4: 'Skill DMG Cap',
-  5: 'C.A. DMG', 6: 'C.A. DMG Cap', 7: 'Stamina', 8: 'Enmity',
-  9: 'Critical Hit', 10: 'Double Attack', 11: 'Triple Attack', 12: 'DEF',
-  13: 'Healing', 14: 'Debuff Resistance', 15: 'Dodge'
-}
-
-// Aetherial (earring) modifier ID to name mapping
-// Note: Element-based stats will use typeName directly (e.g., "Fire ATK Up")
-const AETHERIAL_NAMES = {
-  1: 'Double Attack', 2: 'Triple Attack', 3: 'Element ATK', 4: 'Element Resistance',
-  5: 'Stamina', 6: 'Enmity', 7: 'Supplemental DMG', 8: 'Critical Hit',
-  9: 'Counters on Dodge', 10: 'Counters on DMG'
-}
-
-// Perpetuity Ring bonus ID to name mapping
-const PERPETUITY_NAMES = {
-  1: 'EM Star Cap', 2: 'ATK', 3: 'HP', 4: 'DMG Cap'
-}
-
-/**
- * Stats that use flat values (not percentages) - by display name
- */
-const FLAT_VALUE_STATS = new Set(['ATK', 'HP', 'Stamina', 'Enmity', 'Supplemental DMG'])
-
-/**
- * Format a ring/earring modifier for display
- */
-function formatModifier(mod, nameMap) {
-  if (!mod || !mod.modifier) return null
-
-  // For element-based stats, prefer the actual typeName (e.g., "Fire ATK Up" instead of "Element ATK")
-  let name = nameMap[mod.modifier] || mod.typeName || `Mod ${mod.modifier}`
-  if ((name === 'Element ATK' || name === 'Element Resistance') && mod.typeName) {
-    name = mod.typeName
-  }
-
-  let value = mod.strength
-
-  // ATK and HP get formatted with commas (for Over Mastery rings)
-  if (name === 'ATK' || name === 'HP') {
-    return `${name} +${Number(value).toLocaleString()}`
-  }
-
-  // Flat value stats (no percentage sign)
-  if (FLAT_VALUE_STATS.has(name)) {
-    return `${name} +${value}`
-  }
-
-  // All other stats are percentages
-  return `${name} +${value}%`
-}
-
-/**
- * Format a perpetuity ring bonus for display
- */
-function formatPerpetuitytBonus(bonus) {
-  if (!bonus || !bonus.modifier) return null
-  const name = PERPETUITY_NAMES[bonus.modifier] || bonus.typeName || `Bonus ${bonus.modifier}`
-  const value = bonus.strength
-  // All perpetuity bonuses are flat values or percentages based on type
-  if (bonus.modifier === 1) {
-    // EM Star Cap is a flat value
-    return `${name} +${value}`
-  }
-  if (bonus.modifier === 2 || bonus.modifier === 3) {
-    // ATK and HP are percentages for perpetuity
-    return `${name} +${value}%`
-  }
-  if (bonus.modifier === 4) {
-    // DMG Cap is a percentage
-    return `${name} +${value}%`
-  }
-  return `${name} +${value}`
-}
-
 /**
  * Render character stats detail view
  */
@@ -1090,7 +787,7 @@ function renderCharacterStatsDetail(container, data) {
     if (char.perpetuityBonuses && char.perpetuityBonuses.length > 0) {
       perpetuityHtml = '<div class="char-stats-section"><div class="char-stats-subheader">Perpetuity Bonuses</div>'
       for (const bonus of char.perpetuityBonuses) {
-        const bonusStr = formatPerpetuitytBonus(bonus)
+        const bonusStr = formatPerpetuityBonus(bonus)
         if (bonusStr) {
           perpetuityHtml += `<div class="char-stats-line">${bonusStr}</div>`
         }
@@ -1166,657 +863,6 @@ function renderCharacterStatsDetail(container, data) {
 
   // Update last render time for detecting new items on next render
   lastCharacterStatsRenderTime = renderTime
-}
-
-// Granblue Fantasy CDN for game assets
-const GBF_CDN = 'https://prd-game-a-granbluefantasy.akamaized.net/assets_en/img/sp/assets'
-
-/**
- * Render database detail view (single character/weapon/summon)
- */
-function renderDatabaseDetail(container, dataType, data) {
-  const id = data.id || data.master?.id
-  const name = data.name || data.master?.name || 'Unknown'
-  const element = data.attribute || data.element || data.master?.attribute || data.master?.element
-
-  let imageUrl = ''
-  let fallbackUrl = ''
-  let imageClass = ''
-
-  // Use GBF's CDN since these are new items not yet on our S3
-  // Use startsWith to handle per-item data types (detail_npc_123, etc.)
-  if (dataType.startsWith('detail_npc')) {
-    imageUrl = `${GBF_CDN}/npc/m/${id}_01.jpg`
-    fallbackUrl = `${GBF_CDN}/npc/m/${id}_01_0.jpg`
-    imageClass = 'character-main'
-  } else if (dataType.startsWith('detail_weapon')) {
-    imageUrl = `${GBF_CDN}/weapon/m/${id}.jpg`
-    imageClass = 'weapon-main'
-  } else if (dataType.startsWith('detail_summon')) {
-    imageUrl = `${GBF_CDN}/summon/m/${id}.jpg`
-    imageClass = 'summon-main'
-  }
-
-  const fallbackAttr = fallbackUrl ? `onerror="this.onerror=null; this.src='${fallbackUrl}'"` : ''
-
-  let html = `
-    <div class="database-detail">
-      <div class="database-detail-image ${imageClass}">
-        <img src="${imageUrl}" alt="${name}" ${fallbackAttr}>
-      </div>
-      <div class="database-detail-info">
-  `
-
-  // Add type-specific stats (includes name, element, proficiency as rows)
-  // Proficiency is in specialty_weapon array for all types
-  const proficiencies = data.master?.specialty_weapon || data.specialty_weapon || []
-
-  if (dataType.startsWith('detail_npc')) {
-    html += renderCharacterStats(data, name, id, element, proficiencies)
-  } else if (dataType.startsWith('detail_weapon')) {
-    html += renderWeaponStats(data, name, id, element, proficiencies[0])
-  } else if (dataType.startsWith('detail_summon')) {
-    html += renderSummonStats(data, name, id, element)
-  }
-
-  html += `
-      </div>
-    </div>
-  `
-
-  container.innerHTML = html
-}
-
-/**
- * Render character stats section
- */
-function renderCharacterStats(data, name, id, element, proficiencies = []) {
-  const master = data.master || data
-  const param = data.param || {}
-
-  const ringed = param.has_npcaugment_constant
-
-  // HP/ATK: check param (current stats), then master (base stats), then top-level
-  const minHp = master.default_hp || data.default_hp
-  const maxHp = param.hp || master.max_hp || data.max_hp
-  const minAtk = master.default_attack || data.default_attack
-  const maxAtk = param.attack || master.max_attack || data.max_attack
-  const level = param.level || master.max_level
-
-  let html = '<div class="database-stats">'
-
-  // Name
-  html += `<div class="stat-row"><span class="stat-label">Name</span><span class="stat-value">${name}</span></div>`
-
-  // ID
-  if (id) {
-    html += `<div class="stat-row"><span class="stat-label">ID</span><span class="stat-value">${id}</span></div>`
-  }
-
-  // Series
-  const seriesId = data.series_id || master.series_id
-  if (seriesId && GAME_CHARACTER_SERIES_NAMES[seriesId]) {
-    html += `<div class="stat-row"><span class="stat-label">Series</span><span class="stat-value">${GAME_CHARACTER_SERIES_NAMES[seriesId]}</span></div>`
-  }
-
-  // Element
-  if (element && GAME_ELEMENT_NAMES[element]) {
-    html += `<div class="stat-row"><span class="stat-label">Element</span><span class="stat-value"><img class="stat-icon" src="${getImageUrl(`labels/element/Label_Element_${GAME_ELEMENT_NAMES[element]}.png`)}" alt="${GAME_ELEMENT_NAMES[element]}"></span></div>`
-  }
-
-  // Proficiency (characters can have up to 2)
-  if (proficiencies.length > 0) {
-    const profIcons = proficiencies
-      .filter(p => GAME_PROFICIENCY_NAMES[p])
-      .map(p => `<img class="stat-icon" src="${getImageUrl(`labels/proficiency/Label_Weapon_${GAME_PROFICIENCY_NAMES[p]}.png`)}" alt="${GAME_PROFICIENCY_NAMES[p]}">`)
-      .join('')
-    if (profIcons) {
-      html += `<div class="stat-row"><span class="stat-label">Proficiency</span><span class="stat-value">${profIcons}</span></div>`
-    }
-  }
-
-  // Uncap stars (based on max level)
-  if (level) {
-    html += `<div class="stat-row"><span class="stat-label">Uncap</span><span class="stat-value">${renderCharacterStars(level)}</span></div>`
-  }
-
-  // Min HP
-  if (minHp) {
-    html += `<div class="stat-row"><span class="stat-label">Min HP</span><span class="stat-value">${Number(minHp).toLocaleString()}</span></div>`
-  }
-
-  // Max HP
-  if (maxHp) {
-    html += `<div class="stat-row"><span class="stat-label">Max HP</span><span class="stat-value">${Number(maxHp).toLocaleString()}</span></div>`
-  }
-
-  // Min ATK
-  if (minAtk) {
-    html += `<div class="stat-row"><span class="stat-label">Min ATK</span><span class="stat-value">${Number(minAtk).toLocaleString()}</span></div>`
-  }
-
-  // Max ATK
-  if (maxAtk) {
-    html += `<div class="stat-row"><span class="stat-label">Max ATK</span><span class="stat-value">${Number(maxAtk).toLocaleString()}</span></div>`
-  }
-
-  // Max Level
-  if (level) {
-    html += `<div class="stat-row"><span class="stat-label">Max Level</span><span class="stat-value">${level}</span></div>`
-  }
-
-  // Perpetuity Ring
-  if (ringed) {
-    html += `<div class="stat-row"><span class="stat-label">Perpetuity Ring</span><span class="stat-value">✓</span></div>`
-  }
-
-  // Comment/description
-  const comment = data.comment || master.comment
-  if (comment) {
-    html += `<div class="stat-row stat-comment"><span class="stat-value">${comment}</span></div>`
-  }
-
-  html += '</div>'
-  return html
-}
-
-/**
- * Render weapon stats section
- */
-function renderWeaponStats(data, name, id, element, proficiency) {
-  const master = data.master || data
-  const param = data.param || {}
-
-  // HP/ATK: check param (current stats), then master (base stats), then top-level
-  const minHp = master.default_hp || data.default_hp
-  const maxHp = param.hp || master.max_hp || data.max_hp
-  const minAtk = master.default_attack || data.default_attack
-  const maxAtk = param.attack || master.max_attack || data.max_attack
-  const level = param.level || master.max_level
-
-  let html = '<div class="database-stats">'
-
-  // Name
-  html += `<div class="stat-row"><span class="stat-label">Name</span><span class="stat-value">${name}</span></div>`
-
-  // ID
-  if (id) {
-    html += `<div class="stat-row"><span class="stat-label">ID</span><span class="stat-value">${id}</span></div>`
-  }
-
-  // Series
-  const seriesId = data.series_id || master.series_id
-  if (seriesId && GAME_WEAPON_SERIES_NAMES[seriesId]) {
-    html += `<div class="stat-row"><span class="stat-label">Series</span><span class="stat-value">${GAME_WEAPON_SERIES_NAMES[seriesId]}</span></div>`
-  }
-
-  // Element
-  if (element && GAME_ELEMENT_NAMES[element]) {
-    html += `<div class="stat-row"><span class="stat-label">Element</span><span class="stat-value"><img class="stat-icon" src="${getImageUrl(`labels/element/Label_Element_${GAME_ELEMENT_NAMES[element]}.png`)}" alt="${GAME_ELEMENT_NAMES[element]}"></span></div>`
-  }
-
-  // Proficiency
-  if (proficiency && GAME_PROFICIENCY_NAMES[proficiency]) {
-    html += `<div class="stat-row"><span class="stat-label">Proficiency</span><span class="stat-value"><img class="stat-icon" src="${getImageUrl(`labels/proficiency/Label_Weapon_${GAME_PROFICIENCY_NAMES[proficiency]}.png`)}" alt="${GAME_PROFICIENCY_NAMES[proficiency]}"></span></div>`
-  }
-
-  // Uncap stars (based on max level)
-  if (level) {
-    html += `<div class="stat-row"><span class="stat-label">Uncap</span><span class="stat-value">${renderWeaponStars(level)}</span></div>`
-  }
-
-  // Min HP
-  if (minHp) {
-    html += `<div class="stat-row"><span class="stat-label">Min HP</span><span class="stat-value">${Number(minHp).toLocaleString()}</span></div>`
-  }
-
-  // Max HP
-  if (maxHp) {
-    html += `<div class="stat-row"><span class="stat-label">Max HP</span><span class="stat-value">${Number(maxHp).toLocaleString()}</span></div>`
-  }
-
-  // Min ATK
-  if (minAtk) {
-    html += `<div class="stat-row"><span class="stat-label">Min ATK</span><span class="stat-value">${Number(minAtk).toLocaleString()}</span></div>`
-  }
-
-  // Max ATK
-  if (maxAtk) {
-    html += `<div class="stat-row"><span class="stat-label">Max ATK</span><span class="stat-value">${Number(maxAtk).toLocaleString()}</span></div>`
-  }
-
-  // Max Level
-  if (level) {
-    html += `<div class="stat-row"><span class="stat-label">Max Level</span><span class="stat-value">${level}</span></div>`
-  }
-
-  // Awakening
-  const arousal = param.arousal
-  if (arousal?.is_arousal_weapon) {
-    html += `<div class="stat-row"><span class="stat-label">Awakening</span><span class="stat-value">${arousal.form_name || 'Attack'} Lv.${arousal.level || 1}</span></div>`
-  }
-
-  // Odiant / Befoulment
-  const odiant = param.odiant
-  if (odiant?.is_odiant_weapon) {
-    const befoulSkill = param.augment_skill_info?.[0]?.[0]
-    const showValue = befoulSkill?.show_value || 'Active'
-    html += `<div class="stat-row"><span class="stat-label">Befoulment</span><span class="stat-value">${showValue}</span></div>`
-    html += `<div class="stat-row"><span class="stat-label">Exorcism</span><span class="stat-value">${odiant.exorcision_level || 0}/${odiant.max_exorcision_level || 5}</span></div>`
-  } else {
-    // AX Skills (non-Odiant only)
-    const axSkills = param.augment_skill_info?.[0]
-    if (axSkills && Object.keys(axSkills).length > 0) {
-      const axCount = Object.keys(axSkills).length
-      html += `<div class="stat-row"><span class="stat-label">AX Skills</span><span class="stat-value">${axCount} skill${axCount > 1 ? 's' : ''}</span></div>`
-    }
-  }
-
-  // Comment/description
-  const comment = data.comment || master.comment
-  if (comment) {
-    html += `<div class="stat-row stat-comment"><span class="stat-value">${comment}</span></div>`
-  }
-
-  html += '</div>'
-  return html
-}
-
-/**
- * Render summon stats section
- */
-function renderSummonStats(data, name, id, element) {
-  const master = data.master || data
-  const param = data.param || {}
-
-  // HP/ATK: check param (current stats), then master (base stats), then top-level
-  const minHp = master.default_hp || data.default_hp
-  const maxHp = param.hp || master.max_hp || data.max_hp
-  const minAtk = master.default_attack || data.default_attack
-  const maxAtk = param.attack || master.max_attack || data.max_attack
-  const level = param.level || master.max_level
-
-  let html = '<div class="database-stats">'
-
-  // Name
-  html += `<div class="stat-row"><span class="stat-label">Name</span><span class="stat-value">${name}</span></div>`
-
-  // ID
-  if (id) {
-    html += `<div class="stat-row"><span class="stat-label">ID</span><span class="stat-value">${id}</span></div>`
-  }
-
-  // Series
-  const seriesId = data.series_id || master.series_id
-  if (seriesId && GAME_SUMMON_SERIES_NAMES[seriesId]) {
-    html += `<div class="stat-row"><span class="stat-label">Series</span><span class="stat-value">${GAME_SUMMON_SERIES_NAMES[seriesId]}</span></div>`
-  }
-
-  // Element
-  if (element && GAME_ELEMENT_NAMES[element]) {
-    html += `<div class="stat-row"><span class="stat-label">Element</span><span class="stat-value"><img class="stat-icon" src="${getImageUrl(`labels/element/Label_Element_${GAME_ELEMENT_NAMES[element]}.png`)}" alt="${GAME_ELEMENT_NAMES[element]}"></span></div>`
-  }
-
-  // Uncap stars (based on max level)
-  if (level) {
-    html += `<div class="stat-row"><span class="stat-label">Uncap</span><span class="stat-value">${renderSummonStars(level)}</span></div>`
-  }
-
-  // Min HP
-  if (minHp) {
-    html += `<div class="stat-row"><span class="stat-label">Min HP</span><span class="stat-value">${Number(minHp).toLocaleString()}</span></div>`
-  }
-
-  // Max HP
-  if (maxHp) {
-    html += `<div class="stat-row"><span class="stat-label">Max HP</span><span class="stat-value">${Number(maxHp).toLocaleString()}</span></div>`
-  }
-
-  // Min ATK
-  if (minAtk) {
-    html += `<div class="stat-row"><span class="stat-label">Min ATK</span><span class="stat-value">${Number(minAtk).toLocaleString()}</span></div>`
-  }
-
-  // Max ATK
-  if (maxAtk) {
-    html += `<div class="stat-row"><span class="stat-label">Max ATK</span><span class="stat-value">${Number(maxAtk).toLocaleString()}</span></div>`
-  }
-
-  // Max Level
-  if (level) {
-    html += `<div class="stat-row"><span class="stat-label">Max Level</span><span class="stat-value">${level}</span></div>`
-  }
-
-  // Sub Aura
-  const subAura = data.sub_skill?.name
-  if (subAura) {
-    html += `<div class="stat-row"><span class="stat-label">Sub Aura</span><span class="stat-value">${subAura}</span></div>`
-  }
-
-  // Comment/description
-  const comment = data.comment || master.comment
-  if (comment) {
-    html += `<div class="stat-row stat-comment"><span class="stat-value">${comment}</span></div>`
-  }
-
-  html += '</div>'
-  return html
-}
-
-/**
- * Render uncap stars for characters based on max level
- * - Max level 80: 4 yellow stars
- * - Max level 100: 4 yellow + 1 blue star
- * - Max level 150: 4 yellow + 1 blue + 1 purple star
- * @param {number} maxLevel - The character's max level
- * @returns {string} HTML string for stars
- */
-function renderCharacterStars(maxLevel) {
-  let html = '<span class="stars">'
-
-  // 4 yellow stars for base (max level 80)
-  for (let i = 0; i < 4; i++) {
-    html += '<span class="star filled"></span>'
-  }
-
-  // 1 blue star for FLB (max level 100)
-  if (maxLevel >= 100) {
-    html += '<span class="star flb"></span>'
-  }
-
-  // 1 purple star for ULB (max level 150)
-  if (maxLevel >= 150) {
-    html += '<span class="star ulb"></span>'
-  }
-
-  html += '</span>'
-  return html
-}
-
-/**
- * Render uncap stars for weapons based on max level
- * - Max level 100: 3 yellow stars
- * - Max level 150: 3 yellow + 1 blue star
- * - Max level 200: 3 yellow + 2 blue stars
- * - Max level 250: 3 yellow + 2 blue + 1 purple star
- * @param {number} maxLevel - The weapon's max level
- * @returns {string} HTML string for stars
- */
-function renderWeaponStars(maxLevel) {
-  let html = '<span class="stars">'
-
-  // 3 yellow stars for base (max level 100)
-  for (let i = 0; i < 3; i++) {
-    html += '<span class="star filled"></span>'
-  }
-
-  // 1 blue star for FLB (max level 150)
-  if (maxLevel >= 150) {
-    html += '<span class="star flb"></span>'
-  }
-
-  // 2nd blue star for ULB (max level 200)
-  if (maxLevel >= 200) {
-    html += '<span class="star flb"></span>'
-  }
-
-  // 1 purple star for transcendence (max level 250)
-  if (maxLevel >= 250) {
-    html += '<span class="star ulb"></span>'
-  }
-
-  html += '</span>'
-  return html
-}
-
-/**
- * Render uncap stars for summons based on max level
- * - Max level 100: 3 yellow stars
- * - Max level 150: 3 yellow + 1 blue star
- * - Max level 200: 3 yellow + 2 blue stars
- * - Max level 250: 3 yellow + 2 blue + 1 purple star
- * @param {number} maxLevel - The summon's max level
- * @returns {string} HTML string for stars
- */
-function renderSummonStars(maxLevel) {
-  let html = '<span class="stars">'
-
-  // 3 yellow stars for base (max level 100)
-  for (let i = 0; i < 3; i++) {
-    html += '<span class="star filled"></span>'
-  }
-
-  // 1 blue star for FLB (max level 150)
-  if (maxLevel >= 150) {
-    html += '<span class="star flb"></span>'
-  }
-
-  // 2nd blue star for ULB (max level 200)
-  if (maxLevel >= 200) {
-    html += '<span class="star flb"></span>'
-  }
-
-  // 1 purple star for transcendence (max level 250)
-  if (maxLevel >= 250) {
-    html += '<span class="star ulb"></span>'
-  }
-
-  html += '</span>'
-  return html
-}
-
-/**
- * Extract items from data based on type
- */
-function extractItems(dataType, data) {
-  if (dataType.startsWith('collection_') || dataType.startsWith('list_') || dataType.startsWith('stash_')) {
-    // Paginated collection - data.pages is an object keyed by page number
-    const pages = Object.values(data)
-    return pages.flatMap(page => page.list || [])
-  }
-  if (dataType.startsWith('party_')) {
-    // Characters are at deck.npc, weapons/summons are at deck.pc
-    const deck = data.deck || {}
-    const pc = deck.pc || {}
-    return [
-      ...toArray(deck.npc),
-      ...toArray(pc.weapons),
-      ...toArray(pc.summons)
-    ].filter(Boolean)
-  }
-  // Single detail item
-  return [data]
-}
-
-/**
- * Get image URL for an item using siero-img S3 CDN
- */
-function getItemImageUrl(dataType, item) {
-  const granblueId = item.master?.id || item.param?.id || item.id
-
-  // Use square images for detail view
-  if (dataType.includes('npc') || dataType.includes('character')) {
-    return getImageUrl(`character-square/${granblueId}_01.jpg`)
-  }
-  if (dataType.includes('weapon')) {
-    return getImageUrl(`weapon-square/${granblueId}.jpg`)
-  }
-  if (dataType.includes('summon')) {
-    return getImageUrl(`summon-square/${granblueId}.jpg`)
-  }
-  if (dataType.includes('artifact')) {
-    const artifactId = item.artifact_id || granblueId
-    return getImageUrl(`artifact-square/${artifactId}.jpg`)
-  }
-  return ''
-}
-
-// Hensei element ID to name mapping
-const ELEMENT_NAMES = {
-  1: 'Wind',
-  2: 'Fire',
-  3: 'Water',
-  4: 'Earth',
-  5: 'Dark',
-  6: 'Light'
-}
-
-// Hensei proficiency ID to name mapping
-const PROFICIENCY_NAMES = {
-  1: 'Sabre',
-  2: 'Dagger',
-  3: 'Axe',
-  4: 'Spear',
-  5: 'Bow',
-  6: 'Staff',
-  7: 'Melee',
-  8: 'Harp',
-  9: 'Gun',
-  10: 'Katana'
-}
-
-// Game element ID to name mapping (raw GBF data)
-const GAME_ELEMENT_NAMES = {
-  1: 'Fire',
-  2: 'Water',
-  3: 'Earth',
-  4: 'Wind',
-  5: 'Light',
-  6: 'Dark'
-}
-
-// Game proficiency ID to name mapping (raw GBF data)
-const GAME_PROFICIENCY_NAMES = {
-  1: 'Sabre',
-  2: 'Dagger',
-  3: 'Axe',
-  4: 'Spear',
-  5: 'Bow',
-  6: 'Staff',
-  7: 'Melee',
-  8: 'Harp',
-  9: 'Gun',
-  10: 'Katana'
-}
-
-// Game series_id to name mapping for characters (raw GBF data)
-const GAME_CHARACTER_SERIES_NAMES = {
-  1: 'Summer',
-  2: 'Yukata',
-  3: 'Valentine',
-  4: 'Halloween',
-  5: 'Holiday',
-  6: 'Zodiac',
-  7: 'Grand',
-  8: 'Fantasy',
-  9: 'Collab',
-  10: 'Eternal',
-  11: 'Evoker',
-  12: 'Saint',
-  13: 'Formal'
-}
-
-// Game series_id to name mapping for weapons (raw GBF data)
-const GAME_WEAPON_SERIES_NAMES = {
-  1: 'Seraphic',
-  2: 'Grand',
-  3: 'Dark Opus',
-  4: 'Revenant',
-  5: 'Primal',
-  6: 'Beast',
-  7: 'Regalia',
-  8: 'Omega',
-  9: 'Olden Primal',
-  10: 'Hollowsky',
-  11: 'Xeno',
-  12: 'Rose',
-  13: 'Ultima',
-  14: 'Bahamut',
-  15: 'Epic',
-  16: 'Cosmos',
-  17: 'Superlative',
-  18: 'Vintage',
-  19: 'Class Champion',
-  20: 'Replica',
-  21: 'Relic',
-  22: 'Rusted',
-  23: 'Sephira',
-  24: 'Vyrmament',
-  25: 'Upgrader',
-  26: 'Astral',
-  27: 'Draconic',
-  28: 'Eternal Splendor',
-  29: 'Ancestral',
-  30: 'New World Foundation',
-  31: 'Ennead',
-  32: 'Militis',
-  33: 'Malice',
-  34: 'Menace',
-  35: 'Illustrious',
-  36: 'Proven',
-  37: 'Revans',
-  38: 'World',
-  39: 'Exo',
-  40: 'Draconic Providence',
-  41: 'Celestial',
-  42: 'Omega Rebirth',
-  43: 'Collab',
-  44: 'Destroyer'
-}
-
-// Game series_id to name mapping for summons (raw GBF data)
-const GAME_SUMMON_SERIES_NAMES = {
-  1: 'Providence',
-  2: 'Genesis',
-  3: 'Magna',
-  4: 'Optimus',
-  5: 'Demi Optimus',
-  6: 'Archangel',
-  7: 'Arcarum',
-  8: 'Epic',
-  9: 'Carbuncle',
-  10: 'Dynamis',
-  12: 'Cryptid',
-  13: 'Six Dragons',
-  14: 'Summer',
-  15: 'Yukata',
-  16: 'Holiday',
-  17: 'Collab',
-  18: 'Bellum',
-  19: 'Crest',
-  20: 'Robur'
-}
-
-/**
- * Get element and proficiency labels for an artifact
- */
-function getArtifactLabels(item) {
-  const element = item.attribute || item.element
-  const proficiency = item.kind || item.weapon_kind
-
-  let html = '<div class="list-item-labels">'
-
-  if (element && GAME_ELEMENT_NAMES[element]) {
-    html += `<img class="label-icon" src="${getImageUrl(`labels/element/Label_Element_${GAME_ELEMENT_NAMES[element]}.png`)}" alt="">`
-  }
-
-  if (proficiency && GAME_PROFICIENCY_NAMES[proficiency]) {
-    html += `<img class="label-icon" src="${getImageUrl(`labels/proficiency/Label_Weapon_${GAME_PROFICIENCY_NAMES[proficiency]}.png`)}" alt="">`
-  }
-
-  html += '</div>'
-  return html
-}
-
-/**
- * Get grid class based on data type
- */
-function getGridClass(dataType) {
-  if (dataType.includes('artifact')) return 'artifacts'
-  if (dataType.includes('npc') || dataType.includes('character')) return 'characters'
-  if (dataType.includes('weapon')) return 'weapons'
-  if (dataType.includes('summon')) return 'summons'
-  return ''
 }
 
 /**
@@ -1982,212 +1028,6 @@ async function handleDetailImport() {
       importBtn.disabled = false
       importBtn.textContent = 'Import'
     }
-  }
-}
-
-// ==========================================
-// SYNC FUNCTIONALITY
-// ==========================================
-
-// Store cached data and preview for sync confirmation
-let pendingSyncData = null
-let pendingSyncPreview = null
-
-/**
- * Handle sync button click - fetch preview then show modal
- */
-async function handleDetailSync() {
-  if (!currentDetailDataType) return
-
-  const syncBtn = document.getElementById('detailSync')
-  if (syncBtn) {
-    syncBtn.disabled = true
-    syncBtn.textContent = 'Checking...'
-  }
-
-  try {
-    // Get cached data (all pages)
-    const response = await chrome.runtime.sendMessage({
-      action: 'getCachedData',
-      dataType: currentDetailDataType
-    })
-
-    if (response.error) {
-      showToast('Sync failed: ' + response.error)
-      return
-    }
-
-    pendingSyncData = response.data
-
-    // Call preview endpoint to see what will be deleted
-    const previewResponse = await chrome.runtime.sendMessage({
-      action: 'previewSyncDeletions',
-      data: response.data,
-      dataType: currentDetailDataType
-    })
-
-    if (previewResponse.error) {
-      showToast('Preview failed: ' + previewResponse.error)
-      return
-    }
-
-    pendingSyncPreview = previewResponse
-
-    // Update modal content with preview
-    updateSyncModalContent(previewResponse)
-    showSyncModal()
-  } catch (error) {
-    showToast('Sync failed: ' + error.message)
-  } finally {
-    if (syncBtn) {
-      syncBtn.disabled = false
-      syncBtn.textContent = 'Full Sync'
-    }
-  }
-}
-
-/**
- * Update sync modal content based on preview
- */
-function updateSyncModalContent(preview) {
-  const warningDiv = document.getElementById('syncWarning')
-  const deleteListDiv = document.getElementById('syncDeleteList')
-
-  if (preview.count > 0) {
-    // Show warning and grid of items to delete
-    warningDiv.classList.remove('hidden')
-
-    // Build the grid of items with images
-    const itemsHtml = preview.willDelete.slice(0, 12).map(item => {
-      const name = item.name || `Unknown`
-      const imageUrl = getSyncPreviewImageUrl(item.granblue_id)
-      return `
-        <div class="sync-delete-item" title="${name}">
-          <img src="${imageUrl}" alt="${name}">
-        </div>
-      `
-    }).join('')
-
-    const moreCount = preview.count > 12
-      ? `<p class="more-items">...and ${preview.count - 12} more</p>`
-      : ''
-
-    if (deleteListDiv) {
-      deleteListDiv.innerHTML = `
-        <p class="delete-count">${preview.count} item${preview.count > 1 ? 's' : ''} will be removed:</p>
-        <div class="sync-delete-grid">${itemsHtml}</div>
-        ${moreCount}
-      `
-      deleteListDiv.classList.remove('hidden')
-    }
-  } else {
-    // No deletions, hide warning
-    warningDiv.classList.add('hidden')
-    if (deleteListDiv) {
-      deleteListDiv.innerHTML = '<p class="no-deletions">No items will be removed.</p>'
-      deleteListDiv.classList.remove('hidden')
-    }
-  }
-}
-
-/**
- * Get image URL for sync preview based on current data type
- */
-function getSyncPreviewImageUrl(granblueId) {
-  if (!granblueId) return ''
-
-  // Determine image type based on current detail data type
-  if (currentDetailDataType?.includes('weapon')) {
-    return getImageUrl(`weapon-square/${granblueId}.jpg`)
-  }
-  if (currentDetailDataType?.includes('summon')) {
-    return getImageUrl(`summon-square/${granblueId}.jpg`)
-  }
-  if (currentDetailDataType?.includes('artifact')) {
-    return getImageUrl(`artifact-square/${granblueId}.jpg`)
-  }
-  if (currentDetailDataType?.includes('npc') || currentDetailDataType?.includes('character')) {
-    return getImageUrl(`character-square/${granblueId}_01.jpg`)
-  }
-  return ''
-}
-
-/**
- * Show sync confirmation modal
- */
-function showSyncModal() {
-  const modal = document.getElementById('syncModal')
-  modal?.classList.remove('hidden')
-}
-
-/**
- * Hide sync confirmation modal
- */
-function hideSyncModal() {
-  const modal = document.getElementById('syncModal')
-  modal?.classList.add('hidden')
-  pendingSyncData = null
-  pendingSyncPreview = null
-}
-
-/**
- * Confirm and execute full sync with reconciliation
- */
-async function confirmSync() {
-  hideSyncModal()
-
-  if (!currentDetailDataType || !pendingSyncData) return
-
-  const syncBtn = document.getElementById('detailSync')
-  if (syncBtn) {
-    syncBtn.disabled = true
-    syncBtn.textContent = 'Syncing...'
-  }
-
-  try {
-    // Upload with full sync options (update existing + reconcile deletions)
-    const uploadResponse = await chrome.runtime.sendMessage({
-      action: 'uploadCollectionData',
-      data: pendingSyncData,
-      dataType: currentDetailDataType,
-      updateExisting: true,
-      isFullInventory: true,
-      reconcileDeletions: true
-    })
-
-    if (uploadResponse.error) {
-      showToast(uploadResponse.error)
-    } else {
-      // Show sync results
-      const total = uploadResponse.created + uploadResponse.updated
-      let msg = `Synced ${total} items`
-
-      if (uploadResponse.reconciliation) {
-        const { deleted, orphaned_grid_items } = uploadResponse.reconciliation
-        if (deleted > 0) {
-          msg += `, removed ${deleted}`
-        }
-        if (orphaned_grid_items?.length > 0) {
-          msg += `, ${orphaned_grid_items.length} orphaned`
-        }
-      }
-
-      showToast(msg)
-
-      if (syncBtn) {
-        syncBtn.textContent = 'Synced'
-        syncBtn.classList.add('synced')
-      }
-    }
-  } catch (error) {
-    showToast('Sync failed: ' + error.message)
-  } finally {
-    if (syncBtn && !syncBtn.classList.contains('synced')) {
-      syncBtn.disabled = false
-      syncBtn.textContent = 'Full Sync'
-    }
-    pendingSyncData = null
-    pendingSyncPreview = null
   }
 }
 
@@ -2511,118 +1351,6 @@ async function handleClearCache() {
 
   showTabStatus(activeTab, 'Cache cleared', 'info')
   setTimeout(() => hideTabStatus(activeTab), 2000)
-}
-
-// ==========================================
-// DATA OPERATIONS
-// ==========================================
-
-/**
- * Handle export for a tab
- */
-async function handleExport(tabName) {
-  const dataType = selectedDataTypes[tabName]
-  if (!dataType) {
-    showTabStatus(tabName, 'Please select data to export', 'error')
-    return
-  }
-
-  const exportBtn = document.getElementById(`export${capitalize(tabName)}`)
-  if (exportBtn) exportBtn.disabled = true
-
-  showTabStatus(tabName, 'Preparing export...', 'info')
-
-  try {
-    // Get cached data
-    const response = await chrome.runtime.sendMessage({
-      action: 'getCachedData',
-      dataType
-    })
-
-    if (response.error) {
-      showTabStatus(tabName, response.error, 'error')
-      return
-    }
-
-    // Upload based on data type
-    let uploadResponse
-    if (dataType === 'party') {
-      uploadResponse = await chrome.runtime.sendMessage({
-        action: 'uploadPartyData',
-        data: response.data
-      })
-    } else if (dataType.startsWith('detail_')) {
-      uploadResponse = await chrome.runtime.sendMessage({
-        action: 'uploadDetailData',
-        data: response.data,
-        dataType
-      })
-    } else if (dataType.startsWith('collection_') || dataType.startsWith('list_') || dataType.startsWith('stash_')) {
-      uploadResponse = await chrome.runtime.sendMessage({
-        action: 'uploadCollectionData',
-        data: response.data,
-        dataType,
-        updateExisting: false
-      })
-    } else {
-      showTabStatus(tabName, 'Export not supported for this data type', 'error')
-      return
-    }
-
-    if (uploadResponse.error) {
-      showTabStatus(tabName, uploadResponse.error, 'error')
-    } else if (uploadResponse.url) {
-      chrome.tabs.create({ url: uploadResponse.url })
-      window.close()
-    } else if (uploadResponse.created !== undefined) {
-      const msg = `Imported: ${uploadResponse.created} new, ${uploadResponse.updated} updated, ${uploadResponse.skipped} skipped`
-      showTabStatus(tabName, msg, 'success')
-      setTimeout(() => hideTabStatus(tabName), 4000)
-    } else {
-      showTabStatus(tabName, 'Export successful!', 'success')
-      setTimeout(() => hideTabStatus(tabName), 2000)
-    }
-  } catch (error) {
-    showTabStatus(tabName, 'Export failed: ' + error.message, 'error')
-  } finally {
-    if (exportBtn) exportBtn.disabled = false
-  }
-}
-
-/**
- * Handle copy for a tab
- */
-async function handleCopy(tabName) {
-  const dataType = selectedDataTypes[tabName]
-  if (!dataType) {
-    showTabStatus(tabName, 'Please select data to copy', 'error')
-    return
-  }
-
-  const copyBtn = document.getElementById(`copy${capitalize(tabName)}`)
-  if (copyBtn) copyBtn.disabled = true
-
-  try {
-    const response = await chrome.runtime.sendMessage({
-      action: 'getCachedData',
-      dataType
-    })
-
-    if (response.error) {
-      showTabStatus(tabName, response.error, 'error')
-      return
-    }
-
-    const jsonString = JSON.stringify(response.data, null, 2)
-    await navigator.clipboard.writeText(jsonString)
-
-    showTabStatus(tabName, `${getDataTypeName(dataType)} data copied!`, 'success')
-    setTimeout(() => hideTabStatus(tabName), 2000)
-  } catch (error) {
-    showTabStatus(tabName, 'Copy failed: ' + error.message, 'error')
-  } finally {
-    if (copyBtn) copyBtn.disabled = false
-  }
 }
 
 // ==========================================
